requirements.txt
Flask==2.3.4
google-generativeai==0.1.0
requests==2.28.2
cryptography==40.0.1
speechrecognition==3.10.0
pygame==2.1.3
gtts==2.2.4
numpy==1.24.4
scipy==1.10.1
gradio==3.36.0
transformers==4.30.2
torch==2.0.1
Pillow==9.5.0
